pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["remeo08 <rmeorux5870@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
pillow = "^10.0.0"
selenium = "^4.10.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
